pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "image-converter"
version = "0.1.1"
description = "CLI tool to convert an image file to another format"
dependencies = ["Pillow"]

[project.scripts]
image-converter = "image_converter.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
